reworked the ghost (separated game from engine)

+++ src/game/client/components/menus.h
@@ -401,7 +401,7 @@ public:
 		char m_aFilename[256];
 		char m_aPlayer[MAX_NAME_LENGTH];
 		
-		float m_Time;
+		int m_Time;
 		
 		bool m_Active;
 		int m_ID;
